#87 claimmoderation has been refactored
