+++ admin-web/src/main/java/io/fundrequest/platform/admin/claim/AutoClaimServiceImpl.java
@@ -1,7 +1,6 @@
 package io.fundrequest.platform.admin.claim;
 
 import io.fundrequest.core.request.RequestService;
-import io.fundrequest.core.request.claim.ClaimService;
 import io.fundrequest.core.request.claim.infrastructure.TrustedRepoRepository;
 import org.apache.commons.lang3.tuple.Pair;
 import org.springframework.scheduling.annotation.Scheduled;
@@ -14,12 +13,12 @@ import java.util.stream.Collectors;
 @Service
 public class AutoClaimServiceImpl implements AutoClaimService {
 
-    private ClaimService claimService;
     private TrustedRepoRepository trustedRepoRepository;
+    private ClaimModerationService claimModerationService;
     private RequestService requestService;
 
-    public AutoClaimServiceImpl(ClaimService claimService, TrustedRepoRepository trustedRepoRepository, RequestService requestService) {
-        this.claimService = claimService;
+    public AutoClaimServiceImpl(final ClaimModerationService claimModerationService, TrustedRepoRepository trustedRepoRepository, RequestService requestService) {
+        this.claimModerationService = claimModerationService;
         this.trustedRepoRepository = trustedRepoRepository;
         this.requestService = requestService;
     }
@@ -28,11 +27,11 @@ public class AutoClaimServiceImpl implements AutoClaimService {
     @Transactional
     public void autoApproveTrustedRepos() {
         Set<String> repos = trustedRepoRepository.findAll().stream().map(r -> r.getOwner().toLowerCase()).collect(Collectors.toSet());
-        claimService.listPendingRequestClaims()
-                    .stream()
-                    .map(r -> Pair.of(r, requestService.findRequest(r.getId())))
-                    .filter(r -> repos.contains(r.getRight().getIssueInformation().getOwner().toLowerCase()))
-                    .map(Pair::getLeft)
-                    .forEach(rc -> claimService.approveClaim(rc.getId()));
+        claimModerationService.listPendingRequestClaims()
+                              .stream()
+                              .map(r -> Pair.of(r, requestService.findRequest(r.getId())))
+                              .filter(r -> repos.contains(r.getRight().getIssueInformation().getOwner().toLowerCase()))
+                              .map(Pair::getLeft)
+                              .forEach(rc -> claimModerationService.approveClaim(rc.getId()));
     }
 }
