+++ admin-web/src/main/java/io/fundrequest/platform/admin/claim/ClaimController.java
@@ -1,6 +1,5 @@
 package io.fundrequest.platform.admin.claim;
 
-import io.fundrequest.core.request.claim.ClaimService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -12,30 +11,29 @@ import org.springframework.web.servlet.view.RedirectView;
 @Controller
 public class ClaimController {
 
-    private ClaimService claimService;
+    private ClaimModerationService claimModerationService;
 
-    public ClaimController(ClaimService claimService) {
-        this.claimService = claimService;
+    public ClaimController(final ClaimModerationService claimModerationService) {
+        this.claimModerationService = claimModerationService;
     }
 
     @GetMapping("/claims")
     public ModelAndView showClaimsPage(Model model) {
-        model.addAttribute("pendingClaims", claimService.listPendingRequestClaims());
+        model.addAttribute("pendingClaims", claimModerationService.listPendingRequestClaims());
         return new ModelAndView("claims");
     }
 
     @PostMapping("/claims/approved")
     public ModelAndView approveRequestClaim(@RequestParam("id") Long id) {
-        claimService.approveClaim(id);
+        claimModerationService.approveClaim(id);
         return
                 new ModelAndView(new RedirectView("/claims", true, true, true));
     }
 
     @PostMapping("/claims/declined")
     public ModelAndView declineRequestClaim(@RequestParam Long id) {
-        claimService.declineClaim(id);
+        claimModerationService.declineClaim(id);
         return
                 new ModelAndView(new RedirectView("/claims", true, true, true));
     }
-
 }
